refactor: hotkey
